# Fix `color-*` performance