--- a/.changeset/young-mangos-sin.md
+++ b/.changeset/young-mangos-sin.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `color-*` performance