--- a/lib/reference/functions.js
+++ b/lib/reference/functions.js
@@ -16,7 +16,22 @@ const camelCaseFunctions = new Set([
 
 const mathFunctions = new Set(['calc', 'clamp', 'max', 'min']);
 
+const colorFunctions = new Set([
+	'rgb',
+	'rgba',
+	'hsl',
+	'hsla',
+	'hwb',
+	'lab',
+	'lch',
+	'oklab',
+	'oklch',
+	'color',
+	'color-mix',
+]);
+
 module.exports = {
 	camelCaseFunctions,
 	mathFunctions,
+	colorFunctions,
 };