--- a/lib/reference/functions.js
+++ b/lib/reference/functions.js
@@ -32,6 +32,6 @@ const colorFunctions = new Set([
 
 module.exports = {
 	camelCaseFunctions,
-	mathFunctions,
 	colorFunctions,
+	mathFunctions,
 };