--- a/lib/reference/functions.js
+++ b/lib/reference/functions.js
@@ -14,11 +14,9 @@ const camelCaseFunctions = new Set([
 	'skewY',
 ]);
 
-const mathFunctions = new Set(['calc', 'clamp', 'max', 'min']);
-
 const colorFunctions = new Set([
-	'rgb',
-	'rgba',
+	'color',
+	'color-mix',
 	'hsl',
 	'hsla',
 	'hwb',
@@ -26,9 +24,11 @@ const colorFunctions = new Set([
 	'lch',
 	'oklab',
 	'oklch',
-	'color',
-	'color-mix',
+	'rgb',
+	'rgba',
 ]);
+
+const mathFunctions = new Set(['calc', 'clamp', 'max', 'min']);
 
 module.exports = {
 	camelCaseFunctions,