--- a/lib/rules/color-function-notation/index.js
+++ b/lib/rules/color-function-notation/index.js
@@ -23,8 +23,10 @@ const meta = {
 	fixable: true,
 };
 
-const LEGACY_FUNCS = new Set(['rgba', 'hsla']);
-const LEGACY_NOTATION_FUNCS = new Set(['rgb', 'rgba', 'hsl', 'hsla']);
+const LEGACY_FUNCTION_NAME = /^(?:rgba|hsla)$/i;
+const LEGACY_NOTATION_FUNCTION = /(?:rgba?|hsla?)\(/i;
+const LEGACY_NOTATION_FUNCTION_NAME = /^(?:rgba?|hsla?)$/i;
+const COMMA = /,/;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions, context) => {
@@ -50,6 +52,10 @@ const rule = (primary, secondaryOptions, context) => {
 		const ignoreWithVarInside = optionsMatches(secondaryOptions, 'ignore', 'with-var-inside');
 
 		root.walkDecls((decl) => {
+			if (!LEGACY_NOTATION_FUNCTION.test(decl.value)) return;
+
+			if (primary === 'modern' && !COMMA.test(decl.value)) return;
+
 			let needsFix = false;
 			const parsedValue = valueParser(getDeclarationValue(decl));
 
@@ -64,7 +70,7 @@ const rule = (primary, secondaryOptions, context) => {
 					return;
 				}
 
-				if (!LEGACY_NOTATION_FUNCS.has(value.toLowerCase())) return;
+				if (!LEGACY_NOTATION_FUNCTION_NAME.test(value)) return;
 
 				if (primary === 'modern' && !hasCommas(node)) return;
 
@@ -93,7 +99,7 @@ const rule = (primary, secondaryOptions, context) => {
 					});
 
 					// Remove trailing 'a' from legacy function name
-					if (LEGACY_FUNCS.has(node.value.toLowerCase())) {
+					if (LEGACY_FUNCTION_NAME.test(node.value)) {
 						node.value = node.value.slice(0, -1);
 					}