--- a/lib/rules/color-function-notation/index.js
+++ b/lib/rules/color-function-notation/index.js
@@ -54,7 +54,7 @@ const rule = (primary, secondaryOptions, context) => {
 		root.walkDecls((decl) => {
 			if (!LEGACY_NOTATION_FUNCTION.test(decl.value)) return;
 
-			if (primary === 'modern' && !COMMA.test(decl.value)) return;
+			if (primary === 'modern' && !decl.value.includes(',')) return;
 
 			let needsFix = false;
 			const parsedValue = valueParser(getDeclarationValue(decl));