--- a/lib/rules/color-function-notation/index.js
+++ b/lib/rules/color-function-notation/index.js
@@ -24,7 +24,7 @@ const meta = {
 };
 
 const LEGACY_FUNCTION_NAME = /^(?:rgba|hsla)$/i;
-const LEGACY_NOTATION_FUNCTION = /(?:rgba?|hsla?)\(/i;
+const LEGACY_NOTATION_FUNCTION = /\b(?:rgba?|hsla?)\(/i;
 const LEGACY_NOTATION_FUNCTION_NAME = /^(?:rgba?|hsla?)$/i;
 const COMMA = /,/;
 
@@ -99,8 +99,8 @@ const rule = (primary, secondaryOptions, context) => {
 					});
 
 					// Remove trailing 'a' from legacy function name
-					if (LEGACY_FUNCTION_NAME.test(node.value)) {
-						node.value = node.value.slice(0, -1);
+					if (LEGACY_FUNCTION_NAME.test(value)) {
+						node.value = value.slice(0, -1);
 					}
 
 					needsFix = true;