--- a/lib/rules/color-function-notation/index.js
+++ b/lib/rules/color-function-notation/index.js
@@ -26,7 +26,6 @@ const meta = {
 const LEGACY_FUNCTION_NAME = /^(?:rgba|hsla)$/i;
 const LEGACY_NOTATION_FUNCTION = /\b(?:rgba?|hsla?)\(/i;
 const LEGACY_NOTATION_FUNCTION_NAME = /^(?:rgba?|hsla?)$/i;
-const COMMA = /,/;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions, context) => {