--- a/lib/rules/color-hex-alpha/index.js
+++ b/lib/rules/color-hex-alpha/index.js
@@ -1,6 +1,8 @@
 'use strict';
 
 const declarationValueIndex = require('../../utils/declarationValueIndex');
+const hasValidHex = require('../../utils/hasValidHex');
+const isValidHex = require('../../utils/isValidHex');
 const report = require('../../utils/report');
 const ruleMessages = require('../../utils/ruleMessages');
 const validateOptions = require('../../utils/validateOptions');
@@ -17,8 +19,6 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/color-hex-alpha',
 };
 
-const HEX = /^#(?:[\da-f]{3,4}|[\da-f]{6}|[\da-f]{8})$/i;
-
 /** @type {import('stylelint').Rule} */
 const rule = (primary) => {
 	return (root, result) => {
@@ -30,6 +30,8 @@ const rule = (primary) => {
 		if (!validOptions) return;
 
 		root.walkDecls((decl) => {
+			if (!hasValidHex(decl.value)) return;
+
 			const parsedValue = valueParser(decl.value);
 
 			parsedValue.walk((node) => {
@@ -71,7 +73,7 @@ function isUrlFunction({ type, value }) {
  * @param {import('postcss-value-parser').Node} node
  */
 function isHexColor({ type, value }) {
-	return type === 'word' && HEX.test(value);
+	return type === 'word' && isValidHex(value);
 }
 
 /**