--- a/lib/rules/color-hex-case/index.js
+++ b/lib/rules/color-hex-case/index.js
@@ -21,7 +21,8 @@ const meta = {
 	deprecated: true,
 };
 
-const HEX = /^#[0-9A-Za-z]+/;
+const HEX = /^#[\da-z]+/i;
+const CONTAINS_HEX = /#[\da-z]+/i;
 const IGNORED_FUNCTIONS = new Set(['url']);
 
 /** @type {import('stylelint').Rule} */
@@ -37,6 +38,8 @@ const rule = (primary, _secondaryOptions, context) => {
 		}
 
 		root.walkDecls((decl) => {
+			if (!CONTAINS_HEX.test(decl.value)) return;
+
 			const parsedValue = valueParser(getDeclarationValue(decl));
 			let needsFix = false;