--- a/lib/rules/color-hex-length/index.js
+++ b/lib/rules/color-hex-length/index.js
@@ -20,7 +20,8 @@ const meta = {
 	fixable: true,
 };
 
-const HEX = /^#[0-9A-Za-z]+/;
+const HEX = /^#[\da-z]+$/i;
+const CONTAINS_HEX = /#[\da-z]+/i;
 const IGNORED_FUNCTIONS = new Set(['url']);
 
 /** @type {import('stylelint').Rule} */
@@ -36,6 +37,8 @@ const rule = (primary, _secondaryOptions, context) => {
 		}
 
 		root.walkDecls((decl) => {
+			if (!CONTAINS_HEX.test(decl.value)) return;
+
 			const parsedValue = valueParser(getDeclarationValue(decl));
 			let needsFix = false;