--- a/lib/rules/color-named/index.js
+++ b/lib/rules/color-named/index.js
@@ -11,6 +11,9 @@ const validateOptions = require('../../utils/validateOptions');
 const valueParser = require('postcss-value-parser');
 const { isRegExp, isString } = require('../../utils/validateTypes');
 const { colord } = require('./colordUtils');
+const hasNamedColor = require('../../utils/hasNamedColor');
+const hasValidHex = require('../../utils/hasValidHex');
+const hasColorFunction = require('../../utils/hasColorFunction');
 
 const ruleName = 'color-named';
 
@@ -25,6 +28,8 @@ const meta = {
 
 // Todo tested on case insensitivity
 const NODE_TYPES = new Set(['word', 'function']);
+
+const HAS_GRAY_FUNCTION = /gray\(/i;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions) => {
@@ -57,6 +62,19 @@ const rule = (primary, secondaryOptions) => {
 
 			// Return early if the property is to be ignored
 			if (optionsMatches(secondaryOptions, 'ignoreProperties', decl.prop)) {
+				return;
+			}
+
+			if (primary === 'never' && !hasNamedColor(decl.value)) {
+				return;
+			}
+
+			if (
+				primary === 'always-where-possible' &&
+				!hasValidHex(decl.value) &&
+				!hasColorFunction(decl.value) &&
+				!HAS_GRAY_FUNCTION.test(decl.value)
+			) {
 				return;
 			}