--- a/lib/rules/color-named/index.js
+++ b/lib/rules/color-named/index.js
@@ -14,6 +14,7 @@ const { colord } = require('./colordUtils');
 const hasNamedColor = require('../../utils/hasNamedColor');
 const hasValidHex = require('../../utils/hasValidHex');
 const hasColorFunction = require('../../utils/hasColorFunction');
+const { namedColorsKeywords } = require('../../reference/keywords');
 
 const ruleName = 'color-named';
 
@@ -104,9 +105,7 @@ const rule = (primary, secondaryOptions) => {
 				if (
 					primary === 'never' &&
 					type === 'word' &&
-					/^[a-z]+$/iu.test(value) &&
-					value.toLowerCase() !== 'transparent' &&
-					colord(value).isValid()
+					namedColorsKeywords.has(value.toLowerCase())
 				) {
 					complain(
 						messages.rejected(value),