--- a/lib/rules/color-named/index.js
+++ b/lib/rules/color-named/index.js
@@ -66,7 +66,9 @@ const rule = (primary, secondaryOptions) => {
 				return;
 			}
 
-			if (primary === 'never' && !hasNamedColor(decl.value)) {
+			const { value: declValue } = decl;
+
+			if (primary === 'never' && !hasNamedColor(declValue)) {
 				return;
 			}