--- a/lib/rules/color-named/index.js
+++ b/lib/rules/color-named/index.js
@@ -30,7 +30,7 @@ const meta = {
 // Todo tested on case insensitivity
 const NODE_TYPES = new Set(['word', 'function']);
 
-const HAS_GRAY_FUNCTION = /gray\(/i;
+const HAS_GRAY_FUNCTION = /\bgray\(/i;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions) => {
@@ -74,14 +74,14 @@ const rule = (primary, secondaryOptions) => {
 
 			if (
 				primary === 'always-where-possible' &&
-				!hasValidHex(decl.value) &&
-				!hasColorFunction(decl.value) &&
-				!HAS_GRAY_FUNCTION.test(decl.value)
+				!hasValidHex(declValue) &&
+				!hasColorFunction(declValue) &&
+				!HAS_GRAY_FUNCTION.test(declValue)
 			) {
 				return;
 			}
 
-			valueParser(decl.value).walk((node) => {
+			valueParser(declValue).walk((node) => {
 				const value = node.value;
 				const type = node.type;
 				const sourceIndex = node.sourceIndex;