--- a/lib/rules/color-no-hex/index.js
+++ b/lib/rules/color-no-hex/index.js
@@ -18,7 +18,8 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/color-no-hex',
 };
 
-const HEX = /^#[0-9A-Za-z]+/;
+const HEX = /^#[\da-z]+$/i;
+const CONTAINS_HEX = /#[\da-z]+/i;
 const IGNORED_FUNCTIONS = new Set(['url']);
 
 /** @type {import('stylelint').Rule} */
@@ -31,6 +32,8 @@ const rule = (primary) => {
 		}
 
 		root.walkDecls((decl) => {
+			if (!CONTAINS_HEX.test(decl.value)) return;
+
 			const parsedValue = valueParser(getDeclarationValue(decl));
 
 			parsedValue.walk((node) => {