--- a/lib/rules/color-no-invalid-hex/index.js
+++ b/lib/rules/color-no-invalid-hex/index.js
@@ -18,6 +18,8 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/color-no-invalid-hex',
 };
 
+const CONTAINS_HEX = /#[\da-z]+/i;
+
 /** @type {import('stylelint').Rule} */
 const rule = (primary) => {
 	return (root, result) => {
@@ -32,12 +34,16 @@ const rule = (primary) => {
 				return;
 			}
 
+			if (!CONTAINS_HEX.test(decl.value)) {
+				return;
+			}
+
 			valueParser(decl.value).walk(({ value, type, sourceIndex }) => {
 				if (type === 'function' && value.endsWith('url')) return false;
 
 				if (type !== 'word') return;
 
-				const hexMatch = /^#[0-9A-Za-z]+/.exec(value);
+				const hexMatch = /^#[\da-z]+/i.exec(value);
 
 				if (!hexMatch) return;