--- a/lib/utils/hasColorFunction.js
+++ b/lib/utils/hasColorFunction.js
@@ -1,6 +1,11 @@
 'use strict';
 
-const HAS_COLOR_FUNCTION = /(?:^|[^\w-])(?:rgba?|hsla?|color|hwb|lab|oklab|lch|oklch|color-mix)\(/i;
+const { colorFunctions } = require('../reference/functions');
+
+const HAS_COLOR_FUNCTION = new RegExp(
+	`(?:^|[^\\w-])(?:${[...colorFunctions.values()].join('|')})\\(`,
+	'i',
+);
 
 /**
  * Check if a value contains any standard CSS color function