--- a/lib/utils/hasColorFunction.js
+++ b/lib/utils/hasColorFunction.js
@@ -2,10 +2,7 @@
 
 const { colorFunctions } = require('../reference/functions');
 
-const HAS_COLOR_FUNCTION = new RegExp(
-	`(?:^|[^\\w-])(?:${[...colorFunctions.values()].join('|')})\\(`,
-	'i',
-);
+const HAS_COLOR_FUNCTION = new RegExp(`\\b(?:${[...colorFunctions.values()].join('|')})\\(`, 'i');
 
 /**
  * Check if a value contains any standard CSS color function