--- a/lib/utils/hasEmptyLine.js
+++ b/lib/utils/hasEmptyLine.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const HAS_EMPTY_LINE = /\n[\r\t ]*\n/;
+
 /**
  * Check if a string contains at least one empty line
  *
@@ -9,5 +11,5 @@
 module.exports = function hasEmptyLine(string) {
 	if (string === '' || string === undefined) return false;
 
-	return /\n[\r\t ]*\n/.test(string);
+	return HAS_EMPTY_LINE.test(string);
 };