--- a/lib/utils/hasLessInterpolation.js
+++ b/lib/utils/hasLessInterpolation.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const HAS_LESS_INTERPOLATION = /@\{.+?\}/;
+
 /**
  * Check whether a string has less interpolation
  *
@@ -7,5 +9,5 @@
  * @return {boolean} If `true`, a string has less interpolation
  */
 module.exports = function hasLessInterpolation(string) {
-	return /@\{.+?\}/.test(string);
+	return HAS_LESS_INTERPOLATION.test(string);
 };