--- a/lib/utils/hasNamedColor.js
+++ b/lib/utils/hasNamedColor.js
@@ -2,10 +2,7 @@
 
 const { namedColorsKeywords } = require('../reference/keywords');
 
-const HAS_NAMED_COLOR = new RegExp(
-	`(?:^|[^\\w-])(?:${[...namedColorsKeywords.values()].join('|')})(?:$|[^\\w-])`,
-	'i',
-);
+const HAS_NAMED_COLOR = new RegExp(`\\b(?:${[...namedColorsKeywords.values()].join('|')})\\b`, 'i');
 
 /**
  * Check if a value contains any standard CSS named color