--- a/lib/utils/hasPsvInterpolation.js
+++ b/lib/utils/hasPsvInterpolation.js
@@ -1,10 +1,12 @@
 'use strict';
 
+const HAS_PSV_INTERPOLATION = /\$\(.+?\)/;
+
 /**
  * Check whether a string has postcss-simple-vars interpolation
  *
  * @param {string} string
  */
 module.exports = function hasPsvInterpolation(string) {
-	return /\$\(.+?\)/.test(string);
+	return HAS_PSV_INTERPOLATION.test(string);
 };