--- a/lib/utils/hasScssInterpolation.js
+++ b/lib/utils/hasScssInterpolation.js
@@ -1,10 +1,12 @@
 'use strict';
 
+const HAS_SCSS_INTERPOLATION = /#\{.+?\}/;
+
 /**
  * Check whether a string has scss interpolation
  *
  * @param {string} string
  */
 module.exports = function hasScssInterpolation(string) {
-	return /#\{.+?\}/.test(string);
+	return HAS_SCSS_INTERPOLATION.test(string);
 };