--- a/lib/utils/hasTplInterpolation.js
+++ b/lib/utils/hasTplInterpolation.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const HAS_TPL_INTERPOLATION = /\{.+?\}/;
+
 /**
  * Check whether a string has JS template literal interpolation or HTML-like template
  *
@@ -7,5 +9,5 @@
  * @return {boolean} If `true`, a string has template literal interpolation
  */
 module.exports = function hasTplInterpolation(string) {
-	return /\{.+?\}/.test(string);
+	return HAS_TPL_INTERPOLATION.test(string);
 };