--- a/lib/utils/isStandardSyntaxUrl.js
+++ b/lib/utils/isStandardSyntaxUrl.js
@@ -5,6 +5,9 @@ const hasPsvInterpolation = require('../utils/hasPsvInterpolation');
 const hasScssInterpolation = require('../utils/hasScssInterpolation');
 const hasTplInterpolation = require('../utils/hasTplInterpolation');
 
+const IS_LESS_VARIABLE_IN_URL = /^@@?[\w-]+$/;
+const IS_SCSS_VARIABLE_IN_URL = /^[$\s\w+\-,./*'"]+$/;
+
 /**
  * Check whether a URL is standard
  *
@@ -33,14 +36,14 @@ module.exports = function isStandardSyntaxUrl(url) {
 	// Less variable works only at the beginning
 	// Check is less variable, allow use '@url/some/path'
 	// https://github.com/less/less.js/blob/3.x/lib/less/parser/parser.js#L547
-	if (url.startsWith('@') && /^@@?[\w-]+$/.test(url)) {
+	if (url.startsWith('@') && IS_LESS_VARIABLE_IN_URL.test(url)) {
 		return false;
 	}
 
 	// In url without quotes scss variable can be everywhere
 	// But in this case it is allowed to use only specific characters
 	// Also forbidden "/" at the end of url
-	if (url.includes('$') && /^[$\s\w+\-,./*'"]+$/.test(url) && !url.endsWith('/')) {
+	if (url.includes('$') && IS_SCSS_VARIABLE_IN_URL.test(url) && !url.endsWith('/')) {
 		return false;
 	}