--- a/lib/utils/isValidHex.js
+++ b/lib/utils/isValidHex.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const IS_VALID_HEX = /^#(?:[\da-f]{3,4}|[\da-f]{6}|[\da-f]{8})$/i;
+
 /**
  * Check if a value is a valid 3, 4, 6 or 8 digit hex
  *
@@ -7,5 +9,5 @@
  * @returns {boolean}
  */
 module.exports = function isValidHex(value) {
-	return /^#(?:[0-9a-fA-F]{3,4}|[0-9a-fA-F]{6}|[0-9a-fA-F]{8})$/.test(value);
+	return IS_VALID_HEX.test(value);
 };