--- a/lib/utils/isValidIdentifier.js
+++ b/lib/utils/isValidIdentifier.js
@@ -1,5 +1,10 @@
 'use strict';
 
+const ISO_10646_CHARACTERS = /\\[0-9a-f]{1,6}(\\r\\n|[ \t\r\n\f])?/gi;
+const SINGLE_ESCAPED_CHARACTERS = /\\./g;
+const NON_WORD_OR_DASH = /[^\w-]/;
+const DIGIT = /\d/;
+
 /**
  * Returns whether a string is a valid CSS identifier
  * (i.e. only alphanumeric characters, `-`, and `_`;
@@ -17,18 +22,18 @@ module.exports = function isValidIdentifier(ident) {
 	// trims, removes ISO 10646 characters, and singly-escaped characters
 	const trimmedIdent = ident
 		.trim()
-		.replace(/\\[0-9a-f]{1,6}(\\r\\n|[ \t\r\n\f])?/gi, '')
-		.replace(/\\./g, '');
+		.replace(ISO_10646_CHARACTERS, '')
+		.replace(SINGLE_ESCAPED_CHARACTERS, '');
 
-	if (/[^\w-]/.test(trimmedIdent)) {
+	if (NON_WORD_OR_DASH.test(trimmedIdent)) {
 		return false;
 	}
 
-	if (/\d/.test(trimmedIdent.charAt(0))) {
+	if (DIGIT.test(trimmedIdent.charAt(0))) {
 		return false;
 	}
 
-	if (trimmedIdent.charAt(0) === '-' && /\d/.test(trimmedIdent.charAt(1))) {
+	if (trimmedIdent.charAt(0) === '-' && DIGIT.test(trimmedIdent.charAt(1))) {
 		return false;
 	}