<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See:
- #6853
- #6869

> Is there anything in the PR that needs further explanation?

- added a fast abort for all `color-*` rules
- added some utilities in support of the `color-*` rules
- ensures that regexp's in utilities are declared once and then reused.

----

I don't mind splitting this up into multiple PR's, one for each rule, but the changes are all similar and minor.

I am unsure what the preference is in this project for balancing the extra noise and review work of multiple PR's vs. the clarity of small chunks. I didn't want to open ±10 PR's without prior discussion :)