@Mouvedia 2023-05-28
What about having a `utils/interpolation.js` with 5 function exports?
e.g. `hasLess`, `hasPsv`, `hasScss`, `hasTpl`, `hasInterpolation`
If not, is it a convention to have only one export per file for utils?