@romainmenke 2023-05-29
## `color-hex-alpha`

before :

```bash
git checkout main && npm run benchmark-rule color-hex-alpha always

# Warnings: 12980
# Mean: 101.40357111764703 ms
# Deviation: 13.776733967718119 ms


git checkout main && npm run benchmark-rule color-hex-alpha never 

# Warnings: 0
# Mean: 83.61844911111112 ms
# Deviation: 11.006563626355849 ms
```

after :

```bash
git checkout fix-color-rules-performance--adaptable-asian-elephant-740c875825 && npm run benchmark-rule color-hex-alpha always

# Warnings: 12980
# Mean: 81.74926383333332 ms
# Deviation: 7.977572095601296 ms


git checkout fix-color-rules-performance--adaptable-asian-elephant-740c875825 && npm run benchmark-rule color-hex-alpha never

# Warnings: 0
# Mean: 69.26526377777779 ms
# Deviation: 8.457018103395368 ms
```